feat: add download video functionality

See merge request root/spindle_frontend!14

+++ src/shared/components/video-card/video-card.tsx
@@ -1,4 +1,5 @@
 import React, { useEffect, useRef, useState } from 'react';
+import { useMutation } from 'react-query';
 import { useNavigate } from 'react-router-dom';
 import styled from '@emotion/styled/macro';
 import clsx from 'clsx';
@@ -25,7 +26,10 @@ import { Checkbox } from 'shared/components/checkbox/checkbox';
 import { Icon } from 'shared/components/icon/icon';
 import { ICON_COLLECTION } from 'shared/components/icon/icon-list';
 import { Typography } from 'shared/components/typography/typography';
-import { ActionMenu } from 'shared/components/video-card/action-menu';
+import {
+	ActionMenu,
+	VideoActionMenuProps,
+} from 'shared/components/video-card/action-menu';
 
 const StyledCheckbox = styled(Checkbox)`
 	position: absolute;
@@ -157,25 +161,28 @@ const StyledBadgeIcon = styled(Icon)`
 	margin-left: 5px;
 `;
 
-interface VideoCardProps {
+export interface VideoCardProps {
 	video: IVideo;
+	activeActions?: VideoActionMenuProps['activeActions'];
 	isSelectMode: boolean;
 	className?: string;
 	checked: boolean;
 	onChecked: (IVideo) => void;
-	onDelete: (video: IVideo) => void;
+	onDelete?: (video: IVideo) => void;
 }
 
 export const VideoCard: React.FC<VideoCardProps> = ({
 	checked,
 	isSelectMode,
 	video,
+	activeActions,
 	className,
 	onChecked,
 	onDelete,
 }) => {
 	const navigate = useNavigate();
 	const copyLinkTimerRef = useRef<ReturnType<typeof setTimeout>>();
+	const downloadVideoMutation = useMutation(VideoApi.downloadVideoById);
 
 	const [isLinkCopied, setIsLinkCopied] = useState(false);
 
@@ -201,17 +208,17 @@ export const VideoCard: React.FC<VideoCardProps> = ({
 
 	const handleDownload = (e) => {
 		e.stopPropagation();
-		VideoApi.downloadVideoById(video.id);
+		downloadVideoMutation.mutate({ id: video.id, title: video.title });
 	};
 
 	const handleOpenDeleteModal = (e) => {
 		e.stopPropagation();
-		onDelete(video);
+		onDelete && onDelete(video);
 	};
 
 	const handleCopyLink = async (e) => {
 		e.stopPropagation();
-		if (isLinkCopied) return;
+		if (isLinkCopied || !navigator.clipboard) return;
 
 		await navigator.clipboard.writeText(
 			VIDEO_ROUTES.VIDEO.generateExternalPath(video.id),
@@ -273,6 +280,7 @@ export const VideoCard: React.FC<VideoCardProps> = ({
 				/>
 				<StyledActionMenu
 					isLinkCopied={isLinkCopied}
+					activeActions={activeActions}
 					onDownload={handleDownload}
 					onDelete={handleOpenDeleteModal}
 					onCopyLink={handleCopyLink}
